remove pre-1.18 compat code, update README
